feat(stream): port syslog plugin
