+++ conf/config-default.yaml
@@ -402,6 +402,7 @@ stream_plugins: # sorted by priority
   - ip-restriction                 # priority: 3000
   - limit-conn                     # priority: 1003
   - mqtt-proxy                     # priority: 1000
+  - syslog                         # priority: 401
   # <- recommend to use priority (0, 100) for your custom plugins
 
 #wasm:
